Split details view out, fix checkbox state
